+++ ratarmount.py
@@ -41,7 +41,7 @@ from ratarmountcore import (
 )
 
 
-__version__ = '0.11.2'
+__version__ = '0.11.3'
 
 
 def hasNonEmptySupport() -> bool: